game

+++ Assets/Scripts/Core/Model/PlayerModel.cs
@@ -7,7 +7,7 @@ namespace Game.Core.Model
     public interface IPlayerModel
     {
         Player Player { get; }
-        IReadOnlyReactiveProperty<int> Health { get; }
+        IReadOnlyReactiveProperty<float> Health { get; }
         IReadOnlyReactiveProperty<int> JumpsCount { get; }
         IReadOnlyReactiveProperty<bool> IsGrounded { get; }
     }
@@ -16,12 +16,12 @@ namespace Game.Core.Model
     public class PlayerModel : IPlayerModel
     {
         public Player Player;
-        public ReactiveProperty<int> Health = new ReactiveProperty<int>();
+        public ReactiveProperty<float> Health = new ReactiveProperty<float>();
         public ReactiveProperty<int> JumpsCount = new ReactiveProperty<int>();
         public ReactiveProperty<bool> IsGrounded = new ReactiveProperty<bool>();
 
         Player IPlayerModel.Player => Player;
-        IReadOnlyReactiveProperty<int> IPlayerModel.Health => Health;
+        IReadOnlyReactiveProperty<float> IPlayerModel.Health => Health;
         IReadOnlyReactiveProperty<int> IPlayerModel.JumpsCount => JumpsCount;
         IReadOnlyReactiveProperty<bool> IPlayerModel.IsGrounded => IsGrounded;
